requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
SQLAlchemy==1.4.2
PyJWT==2.0.1
bcrypt==3.2.0
pytest==6.2.2
python-dotenv